Fix injection mode and add a test for it.
Rework finding the location to do the test injection - add a function to find
the correct callsite, determineInjectionCallsite(), which searches for the
second stack frame in user code and uses its source file and position.

Add a basic test of the assertion which executes a file containing a test
that uses 'with http recorded and injected' and assert the resulting file.