Fix then() block firing after caught() when recording.
Splitting out the performRequest() method introduced a regression that would
try to resolve the request as if it was successful due to an issue with the
ordering of promise chaining.

Add a regression test for this error by destroying the response socket and
ensuring that this will be correctly recorded. Previously this casued a bad
dereference statusCode on a non-existent result thus an unchecked exception.