Fix multiple injections.
When we find an injection point using the stack we are given a line number.
The subsequent search for the "with http recorded and injected" assertion is
done by regex which must find the *next* occurrence.

To do so, we set the index into the source text of the test file from which
to perform the search but this was set equal to the line number causing the
code to do multiple replacements as the same occurrence thus file position.

Fix this by converting the callsite line number to an index into the source.