Fix the case where signalling an error raises an exception.
Surround the signalling in a try block which will catch any expection that
occurs as a result. We ignore any error raised and now unconditionally do a
reject() for all failures returned from within the delegated assertion. This
relies on Unexpected 10.15.x+ and its loosened promise checks, but is the
only real solution to handle all the possible cases.