Address the case of the first chunk containing a portion of the body.
This change is a better version of what was attempted in 9.2.1 which sought
to identify the boundary between the header and body of the request. Rather
than fiddling with the trailer chunk we instead locate the byte seqeunce
which acts a separator and mark subsequent bytes as part of the body.