Nudge the recorder in the direction of mocker wrt resolution value.
The mocker records all circumstances, even errors, in its timeline
such that all assertions can be done after the fact.

Start doing this for the recorder - make the value when resolving
uniform and set the timeline and fulfilmentValue on the instance.
Use the opportunity to align naming by doubling down on "timeline".