Make verification work correctly when POSTing a JSON request body.
Given verification mode operates on request descriptions we must
that any JSON request body must be converted to a buffer before
the request is performed against the real system.

Split out the code checking for such a body into an isBodyJson()
function and re-use within the verifier to convert if necessary.