Change unexpected-documentation-site-generator to a devDependency
It was moved from devDeps to optionalDependencies in this commit:
https://github.com/unexpectedjs/unexpected-mitm/commit/1efd196cf5d62e5b0f384e5fc2f753a58f505b26

I tested installing the module on node 6 and did not find any errors, so
it seems safe to "revert" this change.

And optionalDependency is not a suitable replacement for a devDependency
as it will also be installed when consumers install the module. If this
change is not enough, we might need to consider installing it lazily -
as in, only install it just-in-time before invoking the build commands.