Introduce formatRecordedExchange() to contain the serialisation logic.
Previously the Recorder operated on and resolved with plain objects.
Now that the timeline is populated with HttpExchange objects which
are satisfied agaist plain spec objects we need more complete code
to perform the format change. Using the previous code as a guide we
add a new function which does the conversion in a compatible way.

In order to keep written output comparison working it must now
account for the format differences between the timeline values
and what we write out (primarily to do with the url property).