Restore requestDescriptions as a getter on the mocker object.
The refactoring to separate out the mock strategies moved the
requestDescriptions into the strategy object - but some consumers
are dependent on being able to deference it from the mocha. Place
a getter on the mocker to restore compatibility. Note also that
the refactoring was large enough scale to warrant a new minor, so
this commit will be followed up with that.